Move some of the update logic from dashboard to StepUpdater
