+++ apps/step_update/cli.py
@@ -1,6 +1,7 @@
 import difflib
 import sys
 from datetime import datetime
+from enum import Enum
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
 
@@ -19,6 +20,90 @@ log = structlog.get_logger()
 
 # If a new version is not specified, assume current date.
 STEP_VERSION_NEW = datetime.now().strftime("%Y-%m-%d")
+
+
+# Define labels for update states.
+class UpdateState(Enum):
+    UNKNOWN = "Unknown"
+    UP_TO_DATE = "No updates known"
+    OUTDATED = "Outdated"
+    MINOR_UPDATE = "Minor update possible"
+    MAJOR_UPDATE = "Major update possible"
+    ARCHIVABLE = "Archivable"
+
+
+# List of identifiers of steps that should be considered as non-updateable.
+# NOTE: The identifier is the step name without the version (and without the "data://").
+NON_UPDATEABLE_IDENTIFIERS = [
+    # All population-related datasets.
+    "garden/demography/population",
+    "garden/gapminder/population",
+    "garden/hyde/baseline",
+    "garden/un/un_wpp",
+    "meadow/gapminder/population",
+    "meadow/hyde/baseline",
+    "meadow/hyde/general_files",
+    "meadow/un/un_wpp",
+    "open_numbers/open_numbers/gapminder__systema_globalis",
+    "open-numbers/ddf--gapminder--systema_globalis",
+    "snapshot/hyde/general_files.zip",
+    "snapshot/hyde/baseline.zip",
+    "snapshot/gapminder/population.xlsx",
+    "snapshot/un/un_wpp.zip",
+    # Regions dataset.
+    "garden/regions/regions",
+    # Old WB income groups.
+    "garden/wb/wb_income",
+    "meadow/wb/wb_income",
+    "walden/wb/wb_income",
+    # New WB income groups.
+    "garden/wb/income_groups",
+    "meadow/wb/income_groups",
+    "snapshot/wb/income_groups.xlsx",
+    # World Bank country shapes.
+    "snapshot/countries/world_bank.zip",
+    # World Bank WDI.
+    "snapshot/worldbank_wdi/wdi.zip",
+    "meadow/worldbank_wdi/wdi",
+    "garden/worldbank_wdi/wdi",
+    # Other steps we don't want to update (because the underlying data does not get updated).
+    # TODO: We need a better way to achieve this, for example adding update_period_days to all steps and snapshots.
+    #  A simpler alternative would be to move these steps to a separate file in a meaningful place.
+    #  Another option is to have "playlists", e.g. "climate_change_explorer" with the identifiers of steps to update.
+    "meadow/epa/ocean_heat_content",
+    "snapshot/epa/ocean_heat_content_annual_world_700m.csv",
+    "snapshot/epa/ocean_heat_content_annual_world_2000m.csv",
+    "garden/epa/ocean_heat_content",
+    "meadow/epa/ocean_heat_content",
+    "meadow/epa/ice_sheet_mass_balance",
+    "snapshot/epa/ice_sheet_mass_balance.csv",
+    "garden/epa/ice_sheet_mass_balance",
+    "meadow/epa/ice_sheet_mass_balance",
+    "meadow/epa/ghg_concentration",
+    "snapshot/epa/co2_concentration.csv",
+    "snapshot/epa/ch4_concentration.csv",
+    "snapshot/epa/n2o_concentration.csv",
+    "garden/epa/ghg_concentration",
+    "meadow/epa/ghg_concentration",
+    "meadow/epa/mass_balance_us_glaciers",
+    "snapshot/epa/mass_balance_us_glaciers.csv",
+    "garden/epa/mass_balance_us_glaciers",
+    "meadow/epa/mass_balance_us_glaciers",
+    "meadow/climate/antarctic_ice_core_co2_concentration",
+    "snapshot/climate/antarctic_ice_core_co2_concentration.xls",
+    "garden/climate/antarctic_ice_core_co2_concentration",
+    "meadow/climate/antarctic_ice_core_co2_concentration",
+    "meadow/climate/global_sea_level",
+    "snapshot/climate/global_sea_level.csv",
+    "garden/climate/global_sea_level",
+    "meadow/climate/global_sea_level",
+]
+
+# List of dependencies to ignore when calculating the update state.
+# This is done to avoid a certain common dependency (e.g. hyde) to make all steps appear as needing a major update.
+DEPENDENCIES_TO_IGNORE = [
+    "snapshot://hyde/2017/general_files.zip",
+]
 
 
 class StepUpdater:
@@ -45,6 +130,70 @@ class StepUpdater:
         self.steps_df = self.tracker.steps_df.copy()
         # Select only active steps.
         self.steps_df = self.steps_df[self.steps_df["state"] == "active"].reset_index(drop=True)
+        # Add steps update state.
+        self.add_steps_update_state()
+
+    def add_steps_update_state(self) -> None:
+        # To speed up calculations, create a dictionary with all info in steps_df.
+        steps_dict = self.steps_df.set_index("step").to_dict(orient="index")
+
+        # Add a column with the dependencies that are not their latest version.
+        self.steps_df["updateable_dependencies"] = [
+            [
+                dependency
+                for dependency in dependencies
+                if (dependency not in DEPENDENCIES_TO_IGNORE) and (not steps_dict[dependency]["is_latest"])
+            ]
+            for dependencies in self.steps_df["all_active_dependencies"]
+        ]
+
+        # Add a column with the total number of dependencies that are not their latest version.
+        self.steps_df["n_updateable_dependencies"] = [
+            len(dependencies) for dependencies in self.steps_df["updateable_dependencies"]
+        ]
+        # Number of snapshot dependencies that are not their latest version.
+        self.steps_df["n_updateable_snapshot_dependencies"] = [
+            sum(
+                [
+                    not steps_dict[dependency]["is_latest"]
+                    if steps_dict[dependency]["channel"] == "snapshot"
+                    else False
+                    for dependency in dependencies
+                    if dependency not in DEPENDENCIES_TO_IGNORE
+                ]
+            )
+            for dependencies in self.steps_df["all_active_dependencies"]
+        ]
+
+        # Add a column with the update state.
+        # By default, the state is unknown.
+        self.steps_df["update_state"] = UpdateState.UNKNOWN.value
+        # If there is a newer version of the step, it is outdated.
+        self.steps_df.loc[~self.steps_df["is_latest"], "update_state"] = UpdateState.OUTDATED.value
+        # If there are any dependencies that are not their latest version, it needs a minor update.
+        # NOTE: If any of those dependencies is a snapshot, it needs a major update (defined in the following line).
+        self.steps_df.loc[
+            (self.steps_df["is_latest"]) & (self.steps_df["n_updateable_dependencies"] > 0), "update_state"
+        ] = UpdateState.MINOR_UPDATE.value
+        # If there are any snapshot dependencies that are not their latest version, it needs a major update.
+        self.steps_df.loc[
+            (self.steps_df["is_latest"]) & (self.steps_df["n_updateable_snapshot_dependencies"] > 0), "update_state"
+        ] = UpdateState.MAJOR_UPDATE.value
+        # If the step does not need to be updated (i.e. update_period_days = 0) or if all dependencies are up to date,
+        # then the step is up to date (in other words, we are not aware of any possible update).
+        self.steps_df.loc[
+            (self.steps_df["update_period_days"] == 0)
+            | (
+                (self.steps_df["is_latest"])
+                & (self.steps_df["n_updateable_snapshot_dependencies"] == 0)
+                & (self.steps_df["n_updateable_dependencies"] == 0)
+            ),
+            "update_state",
+        ] = UpdateState.UP_TO_DATE.value
+        # If a step has no charts and is not the latest version, it is archivable.
+        self.steps_df.loc[
+            (self.steps_df["n_charts"] == 0) & (~self.steps_df["is_latest"]), "update_state"
+        ] = UpdateState.ARCHIVABLE.value
 
     def check_that_step_exists(self, step: str) -> None:
         """Check that step to be updated exists in the active dag."""
