+++ apps/wizard/pages/dashboard.py
@@ -3,7 +3,6 @@
 """
 import subprocess
 from datetime import datetime
-from enum import Enum
 
 import pandas as pd
 import streamlit as st
@@ -12,7 +11,7 @@ from st_aggrid.grid_options_builder import GridOptionsBuilder
 from st_pages import add_indentation
 from structlog import get_logger
 
-from apps.step_update.cli import StepUpdater
+from apps.step_update.cli import NON_UPDATEABLE_IDENTIFIERS, StepUpdater, UpdateState
 from etl.config import ADMIN_HOST, ENV_IS_REMOTE
 from etl.db import can_connect
 
@@ -35,79 +34,6 @@ if not GRAPHER_DATASET_BASE_URL.startswith(("http://", "https://")):
     # Therefore, prepend "http://" (not "https://" because the site is not secure, and the browser will block the link).
     GRAPHER_DATASET_BASE_URL = f"http://{GRAPHER_DATASET_BASE_URL}"
 
-# List of identifiers of steps that should be considered as non-updateable.
-# NOTE: The identifier is the step name without the version (and without the "data://").
-NON_UPDATEABLE_IDENTIFIERS = [
-    # All population-related datasets.
-    "garden/demography/population",
-    "garden/gapminder/population",
-    "garden/hyde/baseline",
-    "garden/un/un_wpp",
-    "meadow/gapminder/population",
-    "meadow/hyde/baseline",
-    "meadow/hyde/general_files",
-    "meadow/un/un_wpp",
-    "open_numbers/open_numbers/gapminder__systema_globalis",
-    "open-numbers/ddf--gapminder--systema_globalis",
-    "snapshot/hyde/general_files.zip",
-    "snapshot/hyde/baseline.zip",
-    "snapshot/gapminder/population.xlsx",
-    "snapshot/un/un_wpp.zip",
-    # Regions dataset.
-    "garden/regions/regions",
-    # Old WB income groups.
-    "garden/wb/wb_income",
-    "meadow/wb/wb_income",
-    "walden/wb/wb_income",
-    # New WB income groups.
-    "garden/wb/income_groups",
-    "meadow/wb/income_groups",
-    "snapshot/wb/income_groups.xlsx",
-    # World Bank country shapes.
-    "snapshot/countries/world_bank.zip",
-    # World Bank WDI.
-    "snapshot/worldbank_wdi/wdi.zip",
-    "meadow/worldbank_wdi/wdi",
-    "garden/worldbank_wdi/wdi",
-    # Other steps we don't want to update (because the underlying data does not get updated).
-    # TODO: We need a better way to achieve this, for example adding update_period_days to all steps and snapshots.
-    #  A simpler alternative would be to move these steps to a separate file in a meaningful place.
-    #  Another option is to have "playlists", e.g. "climate_change_explorer" with the identifiers of steps to update.
-    "meadow/epa/ocean_heat_content",
-    "snapshot/epa/ocean_heat_content_annual_world_700m.csv",
-    "snapshot/epa/ocean_heat_content_annual_world_2000m.csv",
-    "garden/epa/ocean_heat_content",
-    "meadow/epa/ocean_heat_content",
-    "meadow/epa/ice_sheet_mass_balance",
-    "snapshot/epa/ice_sheet_mass_balance.csv",
-    "garden/epa/ice_sheet_mass_balance",
-    "meadow/epa/ice_sheet_mass_balance",
-    "meadow/epa/ghg_concentration",
-    "snapshot/epa/co2_concentration.csv",
-    "snapshot/epa/ch4_concentration.csv",
-    "snapshot/epa/n2o_concentration.csv",
-    "garden/epa/ghg_concentration",
-    "meadow/epa/ghg_concentration",
-    "meadow/epa/mass_balance_us_glaciers",
-    "snapshot/epa/mass_balance_us_glaciers.csv",
-    "garden/epa/mass_balance_us_glaciers",
-    "meadow/epa/mass_balance_us_glaciers",
-    "meadow/climate/antarctic_ice_core_co2_concentration",
-    "snapshot/climate/antarctic_ice_core_co2_concentration.xls",
-    "garden/climate/antarctic_ice_core_co2_concentration",
-    "meadow/climate/antarctic_ice_core_co2_concentration",
-    "meadow/climate/global_sea_level",
-    "snapshot/climate/global_sea_level.csv",
-    "garden/climate/global_sea_level",
-    "meadow/climate/global_sea_level",
-]
-
-# List of dependencies to ignore when calculating the update state.
-# This is done to avoid a certain common dependency (e.g. hyde) to make all steps appear as needing a major update.
-DEPENDENCIES_TO_IGNORE = [
-    "snapshot://hyde/2017/general_files.zip",
-]
-
 # Initialise session state
 ## Selected steps
 st.session_state.selected_steps = st.session_state.get("selected_steps", [])
@@ -120,16 +46,6 @@ if "reload_key" not in st.session_state:
 
 # Logging
 log = get_logger()
-
-
-# Define labels for update states.
-class UpdateState(Enum):
-    UNKNOWN = "Unknown"
-    UP_TO_DATE = "No updates known"
-    OUTDATED = "Outdated"
-    MINOR_UPDATE = "Minor update possible"
-    MAJOR_UPDATE = "Major update possible"
-    ARCHIVABLE = "Archivable"
 
 
 ########################################
@@ -203,14 +119,11 @@ def load_steps_df(reload_key: int) -> pd.DataFrame:
     # Load steps dataframe.
     steps_df = StepUpdater().steps_df
 
-    # To speed up calculations, create a dictionary with all info in steps_df.
-    steps_dict = steps_df.set_index("step").to_dict(orient="index")
-
     # Fix some columns.
     steps_df["full_path_to_script"] = steps_df["full_path_to_script"].fillna("").astype(str)
     steps_df["dag_file_path"] = steps_df["dag_file_path"].fillna("").astype(str)
 
-    # For convenience, convert days to update to a string, and fill nans with "Unknown".
+    # For convenience, convert days to an arbitrarily big number.
     # Otherwise when sorting, nans are placed before negative numbers, and hence it's not easy to first see steps that
     # need to be updated more urgently.
     steps_df["days_to_update"] = steps_df["days_to_update"].fillna(9999)
@@ -228,58 +141,6 @@ def load_steps_df(reload_key: int) -> pd.DataFrame:
     ]
 
     steps_df = steps_df.drop(columns=["db_dataset_name", "db_dataset_id"], errors="raise")
-
-    # Add a column with the dependencies that are not their latest version.
-    steps_df["updateable_dependencies"] = [
-        [
-            dependency
-            for dependency in dependencies
-            if (dependency not in DEPENDENCIES_TO_IGNORE) and (not steps_dict[dependency]["is_latest"])
-        ]
-        for dependencies in steps_df["all_active_dependencies"]
-    ]
-
-    # Add a column with the total number of dependencies that are not their latest version.
-    steps_df["n_updateable_dependencies"] = [len(dependencies) for dependencies in steps_df["updateable_dependencies"]]
-    # Number of snapshot dependencies that are not their latest version.
-    steps_df["n_updateable_snapshot_dependencies"] = [
-        sum(
-            [
-                not steps_dict[dependency]["is_latest"] if steps_dict[dependency]["channel"] == "snapshot" else False
-                for dependency in dependencies
-                if dependency not in DEPENDENCIES_TO_IGNORE
-            ]
-        )
-        for dependencies in steps_df["all_active_dependencies"]
-    ]
-
-    # Add a column with the update state.
-    # By default, the state is unknown.
-    steps_df["update_state"] = UpdateState.UNKNOWN.value
-    # If there is a newer version of the step, it is outdated.
-    steps_df.loc[~steps_df["is_latest"], "update_state"] = UpdateState.OUTDATED.value
-    # If there are any dependencies that are not their latest version, it needs a minor update.
-    # NOTE: If any of those dependencies is a snapshot, it needs a major update (defined in the following line).
-    steps_df.loc[
-        (steps_df["is_latest"]) & (steps_df["n_updateable_dependencies"] > 0), "update_state"
-    ] = UpdateState.MINOR_UPDATE.value
-    # If there are any snapshot dependencies that are not their latest version, it needs a major update.
-    steps_df.loc[
-        (steps_df["is_latest"]) & (steps_df["n_updateable_snapshot_dependencies"] > 0), "update_state"
-    ] = UpdateState.MAJOR_UPDATE.value
-    # If the step does not need to be updated (i.e. update_period_days = 0) or if all dependencies are up to date,
-    # then the step is up to date (in other words, we are not aware of any possible update).
-    steps_df.loc[
-        (steps_df["update_period_days"] == 0)
-        | (
-            (steps_df["is_latest"])
-            & (steps_df["n_updateable_snapshot_dependencies"] == 0)
-            & (steps_df["n_updateable_dependencies"] == 0)
-        ),
-        "update_state",
-    ] = UpdateState.UP_TO_DATE.value
-    # If a step has no charts and is not the latest version, it is archivable.
-    steps_df.loc[(steps_df["n_charts"] == 0) & (~steps_df["is_latest"]), "update_state"] = UpdateState.ARCHIVABLE.value
 
     return steps_df
